Split array.rs into multiple files
