+++ crates/backend-uzu/src/array/common.rs
@@ -4,7 +4,8 @@ use ndarray::{ArrayView, Dimension, IxDyn};
 
 use crate::{
     ArrayElement, DataType,
-    backends::common::{Backend, Buffer, Context, DenseBuffer},
+    array::size_for_shape,
+    backends::common::{Backend, Buffer, DenseBuffer},
 };
 
 #[derive(Debug)]
@@ -197,83 +198,5 @@ impl<B: Backend> Clone for Array<B> {
             shape: self.shape.clone(),
             data_type: self.data_type,
         }
-    }
-}
-
-pub fn size_for_shape(
-    shape: &[usize],
-    data_type: DataType,
-) -> usize {
-    let Some(last_dim) = shape.last() else {
-        return data_type.size_in_bytes();
-    };
-
-    let bits_per_row = last_dim * data_type.size_in_bits();
-    let padded_bytes_per_row = bits_per_row.div_ceil(8);
-
-    let num_rows: usize = shape.iter().rev().skip(1).product();
-
-    num_rows * padded_bytes_per_row
-}
-
-// Array extends Context with helper functions to create arrays from context
-
-pub trait ArrayContextExt {
-    type Backend: Backend;
-
-    fn create_array_uninitialized(
-        &self,
-        shape: &[usize],
-        data_type: DataType,
-        label: &str,
-    ) -> Array<Self::Backend>;
-
-    fn create_array_zeros(
-        &self,
-        shape: &[usize],
-        data_type: DataType,
-        label: &str,
-    ) -> Array<Self::Backend> {
-        let mut array = self.create_array_uninitialized(shape, data_type, label);
-        array.as_bytes_mut().fill(0);
-        array
-    }
-
-    fn create_array_from<T: ArrayElement>(
-        &self,
-        shape: &[usize],
-        data: &[T],
-        label: &str,
-    ) -> Array<Self::Backend> {
-        let size_from_shape: usize = shape.iter().product();
-        assert_eq!(
-            data.len(),
-            size_from_shape,
-            "Shape size {} and data size {} are not equal",
-            size_from_shape,
-            data.len()
-        );
-
-        let mut array = self.create_array_uninitialized(shape, T::data_type(), label);
-        array.as_slice_mut().copy_from_slice(data);
-        array
-    }
-}
-
-impl<C: Context> ArrayContextExt for C {
-    type Backend = C::Backend;
-
-    fn create_array_uninitialized(
-        &self,
-        shape: &[usize],
-        data_type: DataType,
-        label: &str,
-    ) -> Array<Self::Backend> {
-        let buffer_size_bytes = size_for_shape(shape, data_type);
-
-        let mut buffer = self.create_buffer(buffer_size_bytes).expect("Failed to create buffer");
-        buffer.set_label(Some(label));
-
-        unsafe { Array::from_parts(Rc::new(RefCell::new(buffer)), 0, shape, data_type) }
     }
 }
