+++ crates/backend-uzu/src/array/mod.rs
@@ -1,0 +1,7 @@
+mod common;
+mod context;
+mod util;
+
+pub use common::Array;
+pub use context::ArrayContextExt;
+pub use util::size_for_shape;
